rename and refactor class Ui to command package
